pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tellor-supply-analytics"
version = "0.1.0"
description = "Blockchain analytics tool for collecting Tellor token supply data from multiple sources"
readme = "README.md"
requires-python = ">=3.8.1"
license = {text = "MIT"}
authors = [
    {name = "Blockchain Backend Engineering Team", email = "team@tellor.io"}
]
keywords = ["blockchain", "tellor", "analytics", "supply", "ethereum"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Office/Business :: Financial",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

dependencies = [
    "web3>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "discord-webhook>=1.3.0",
    "fastapi>=0.116.0",
    "uvicorn>=0.33.0",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
tellor-supply-analytics = "tellor_supply_analytics.supply_collector:main"

[project.urls]
Homepage = "https://github.com/tellor-io/tellor-supply-analytics"
Repository = "https://github.com/tellor-io/tellor-supply-analytics"
Documentation = "https://github.com/tellor-io/tellor-supply-analytics#readme"
Issues = "https://github.com/tellor-io/tellor-supply-analytics/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/tellor_supply_analytics"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/.venv",
    "/__pycache__",
    "*.pyc",
    "*.pyo",
    "*.pyd",
    ".DS_Store",
]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true 
